refactor(common): move numeric helpers into primitives folder
