+++ crates/tsz-common/src/lib.rs
@@ -40,4 +40,5 @@ pub mod diagnostics;
 // Compiler options for type checking (shared by solver and checker)
 pub mod checker_options;
 pub use checker_options::CheckerOptions;
-pub mod numeric;
+pub mod primitives;
+pub use primitives::numeric;
